+++ src/app/api/lookups/lookups.routes.ts
@@ -1,6 +1,6 @@
 import { Router, Get } from '@lib/restful';
 import { identity } from '@shared/identity';
-import { Result, Responses } from '@core/helpers';
+import { Responses } from '@core/response';
 
 // tslint:disable-next-line: max-line-length
 const nationalities = [{ key: 'Afghan', value: 'أفغاني' }, { key: 'Albanian', value: 'ألباني' }, { key: 'Algerian', value: 'جزائري' }, { key: 'American', value: 'أمريكي' }, { key: 'Andorran', value: 'أندوري' }, { key: 'Angolan', value: 'أنغولي' }, { key: 'Antiguans', value: 'انتيغوا' }, { key: 'Argentinean', value: 'أرجنتيني' }, { key: 'Armenian', value: 'أرميني' }, { key: 'Australian', value: 'أسترالي' }, { key: 'Austrian', value: 'نمساوي' }, { key: 'Azerbaijani', value: 'أذربيجاني' }, { key: 'Bahamian', value: 'باهامى' }, { key: 'Bahraini', value: 'بحريني' }, { key: 'Bangladeshi', value: 'بنجلاديشي' }, { key: 'Barbadian', value: 'باربادوسي' }, { key: 'Barbudans', value: 'بربودا' }, { key: 'Batswana', value: 'بوتسواني' }, { key: 'Belarusian', value: 'بيلاروسي' }, { key: 'Belgian', value: 'بلجيكي' }, { key: 'Belizean', value: 'بليزي' }, { key: 'Beninese', value: 'بنيني' }, { key: 'Bhutanese', value: 'بوتاني' }, { key: 'Bolivian', value: 'بوليفي' }, { key: 'Bosnian', value: 'بوسني' }, { key: 'Brazilian', value: 'برازيلي' }, { key: 'British', value: 'بريطاني' }, { key: 'Bruneian', value: 'بروناى' }, { key: 'Bulgarian', value: 'بلغاري' }, { key: 'Burkinabe', value: 'بوركيني' }, { key: 'Burmese', value: 'بورمي' }, { key: 'Burundian', value: 'بوروندي' }, { key: 'Cambodian', value: 'كمبودي' }, { key: 'Cameroonian', value: 'كاميروني' }, { key: 'Canadian', value: 'كندي' }, { key: 'Cape Verdean', value: 'االرأس الأخضر' }, { key: 'Central African', value: 'وسط أفريقيا' }, { key: 'Chadian', value: 'تشادي' }, { key: 'Chilean', value: 'شيلي' }, { key: 'Chinese', value: 'صينى' }, { key: 'Colombian', value: 'كولومبي' }, { key: 'Comoran', value: 'جزر القمر' }, { key: 'Congolese', value: 'كونغولي' }, { key: 'Costa Rican', value: 'كوستاريكي' }, { key: 'Croatian', value: 'كرواتية' }, { key: 'Cuban', value: 'كوبي' }, { key: 'Cypriot', value: 'قبرصي' }, { key: 'Czech', value: 'تشيكي' }, { key: 'Danish', value: 'دانماركي' }, { key: 'Djibouti', value: 'جيبوتي' }, { key: 'Dominican', value: 'دومينيكاني' }, { key: 'Dutch', value: 'هولندي' }, { key: 'East Timorese', value: 'تيموري شرقي' }, { key: 'Ecuadorean', value: 'اكوادوري' }, { key: 'Egyptian', value: 'مصري' }, { key: 'Emirian', value: 'إماراتي' }, { key: 'Equatorial Guinean', value: 'غيني  استوائي' }, { key: 'Eritrean', value: 'إريتري' }, { key: 'Estonian', value: 'إستوني' }, { key: 'Ethiopian', value: 'حبشي' }, { key: 'Fijian', value: 'فيجي' }, { key: 'Filipino', value: 'فلبيني' }, { key: 'Finnish', value: 'فنلندي' }, { key: 'French', value: 'فرنسي' }, { key: 'Gabonese', value: 'جابوني' }, { key: 'Gambian', value: 'غامبيي' }, { key: 'Georgian', value: 'جورجي' }, { key: 'German', value: 'ألماني' }, { key: 'Ghanaian', value: 'غاني' }, { key: 'Greek', value: 'إغريقي' }, { key: 'Grenadian', value: 'جرينادي' }, { key: 'Guatemalan', value: 'غواتيمالي' }, { key: 'Guinea - Bissauan', value: 'غيني بيساوي' }, { key: 'Guinean', value: 'غيني' }, { key: 'Guyanese', value: 'جوياني' }, { key: 'Haitian', value: 'هايتي' }, { key: 'Herzegovinian', value: 'هرسكي' }, { key: 'Honduran', value: 'هندوراسي' }, { key: 'Hungarian', value: 'هنغاري' }, { key: 'Icelander', value: 'إيسلندي' }, { key: 'Indian', value: 'هندي' }, { key: 'Indonesian', value: 'إندونيسي' }, { key: 'Iranian', value: 'إيراني' }, { key: 'Iraqi', value: 'عراقي' }, { key: 'Irish', value: 'إيرلندي' }, { key: 'Italian', value: 'إيطالي' }, { key: 'Ivorian', value: 'إفواري' }, { key: 'Jamaican', value: 'جامايكي' }, { key: 'Japanese', value: 'ياباني' }, { key: 'Jordanian', value: 'أردني' }, { key: 'Kazakhstani', value: 'كازاخستاني' }, { key: 'Kenyan', value: 'كيني' }, { key: 'Kittian and Nevisian', value: 'كيتياني ونيفيسياني' }, { key: 'Kuwaiti', value: 'كويتي' }, { key: 'Kyrgyz', value: 'قيرغيزستان' }, { key: 'Laotian', value: 'لاوسي' }, { key: 'Latvian', value: 'لاتفي' }, { key: 'Lebanese', value: 'لبناني' }, { key: 'Liberian', value: 'ليبيري' }, { key: 'Libyan', value: 'ليبي' }, { key: 'Liechtensteiner', value: 'ليختنشتايني' }, { key: 'Lithuanian', value: 'لتواني' }, { key: 'Luxembourger', value: 'لكسمبرغي' }, { key: 'Macedonian', value: 'مقدوني' }, { key: 'Malagasy', value: 'مدغشقري' }, { key: 'Malawian', value: 'مالاوى' }, { key: 'Malaysian', value: 'ماليزي' }, { key: 'Maldivan', value: 'مالديفي' }, { key: 'Malian', value: 'مالي' }, { key: 'Maltese', value: 'مالطي' }, { key: 'Marshallese', value: 'مارشالي' }, { key: 'Mauritanian', value: 'موريتاني' }, { key: 'Mauritian', value: 'موريشيوسي' }, { key: 'Mexican', value: 'مكسيكي' }, { key: 'Micronesian', value: 'ميكرونيزي' }, { key: 'Moldovan', value: 'مولدوفي' }, { key: 'Monacan', value: 'موناكو' }, { key: 'Mongolian', value: 'منغولي' }, { key: 'Moroccan', value: 'مغربي' }, { key: 'Mosotho', value: 'ليسوتو' }, { key: 'Motswana', value: 'لتسواني' }, { key: 'Mozambican', value: 'موزمبيقي' }, { key: 'Namibian', value: 'ناميبي' }, { key: 'Nauruan', value: 'ناورو' }, { key: 'Nepalese', value: 'نيبالي' }, { key: 'New Zealander', value: 'نيوزيلندي' }, { key: 'Ni - Vanuatu', value: 'ني فانواتي' }, { key: 'Nicaraguan', value: 'نيكاراغوا' }, { key: 'Nigerien', value: 'نيجري' }, { key: 'North Korean', value: 'كوري شمالي' }, { key: 'Northern Irish', value: 'إيرلندي شمالي' }, { key: 'Norwegian', value: 'نرويجي' }, { key: 'Omani', value: 'عماني' }, { key: 'Pakistani', value: 'باكستاني' }, { key: 'Palauan', value: 'بالاوي' }, { key: 'Palestinian', value: 'فلسطيني' }, { key: 'Panamanian', value: 'بنمي' }, { key: 'Papua New Guinean', value: 'بابوا غينيا الجديدة' }, { key: 'Paraguayan', value: 'باراغواياني' }, { key: 'Peruvian', value: 'بيروفي' }, { key: 'Polish', value: 'بولندي' }, { key: 'Portuguese', value: 'برتغالي' }, { key: 'Qatari', value: 'قطري' }, { key: 'Romanian', value: 'روماني' }, { key: 'Russian', value: 'روسي' }, { key: 'Rwandan', value: 'رواندي' }, { key: 'Saint Lucian', value: 'لوسياني' }, { key: 'Salvadoran', value: 'سلفادوري' }, { key: 'Samoan', value: 'ساموايان' }, { key: 'San Marinese', value: 'سان مارينيز' }, { key: 'Sao Tomean', value: 'ساو توميان' }, { key: 'Saudi', value: 'سعودي' }, { key: 'Scottish', value: 'سكتلندي' }, { key: 'Senegalese', value: 'سنغالي' }, { key: 'Serbian', value: 'صربي' }, { key: 'Seychellois', value: 'سيشلي' }, { key: 'Sierra Leonean', value: 'سيرا ليوني' }, { key: 'Singaporean', value: 'سنغافوري' }, { key: 'Slovakian', value: 'سلوفاكي' }, { key: 'Slovenian', value: 'سلوفيني' }, { key: 'Solomon Islander', value: 'جزر سليمان' }, { key: 'Somali', value: 'صومالي' }, { key: 'South African', value: 'جنوب افريقيي' }, { key: 'South Korean', value: 'كوري جنوبي' }, { key: 'Spanish', value: 'إسباني' }, { key: 'Sri Lankan', value: 'سري لانكي' }, { key: 'Sudanese', value: 'سوداني' }, { key: 'Surinamer', value: 'سورينامي' }, { key: 'Swazi', value: 'سوازي' }, { key: 'Swedish', value: 'سويدي' }, { key: 'Swiss', value: 'سويسري' }, { key: 'Syrian', value: 'سوري' }, { key: 'Taiwanese', value: 'تايواني' }, { key: 'Tajik', value: 'طاجيكي' }, { key: 'Tanzanian', value: 'تنزاني' }, { key: 'Thai', value: 'التايلاندي' }, { key: 'Togolese', value: 'توغواني' }, { key: 'Tongan', value: 'تونجاني' }, { key: 'Trinidadian or Tobagonian', value: 'ترينيدادي أو توباغوني' }, { key: 'Tunisian', value: 'تونسي' }, { key: 'Turkish', value: 'تركي' }, { key: 'Tuvaluan', value: 'توفالي' }, { key: 'Ugandan', value: 'أوغندي' }, { key: 'Ukrainian', value: 'أوكراني' }, { key: 'Uruguayan', value: 'أوروجواي' }, { key: 'Uzbekistani', value: 'أوزبكستاني' }, { key: 'Venezuelan', value: 'فنزويلي' }, { key: 'Vietnamese', value: 'فيتنامي' }, { key: 'Welsh', value: 'ويلزي' }, { key: 'Yemenite', value: 'يمني' }, { key: 'Zambian', value: 'زامبي' }, { key: 'Zimbabwean', value: 'زيمبابوي' }];
