+++ src/app/api/users/users.spec.ts
@@ -2,7 +2,7 @@ import { generatePhoneNumber, generateUsername, UserFixture, getUri, prepareUser
 import { NetworkStatus, Constants } from '@core/helpers';
 import * as faker from 'faker';
 import { Roles } from '@shared/identity';
-const USER_SESSION_ENDPOINT = getUri(`${Constants.Endpoints.USERS}/${Constants.Endpoints.SEARCH}`);
+const USER_SESSION_ENDPOINT = getUri(`${ Constants.Endpoints.USERS }/${ Constants.Endpoints.SEARCH }`);
 
 xdescribe('[INTERGRATION]', () => {
 
@@ -98,22 +98,22 @@ xdescribe('[INTERGRATION]', () => {
             });
         });
 
-        test('should be unverified by default', async () => {
-            const { body: { data: { id } } } = await userFixture.createUser({ verified: null });
-            const { headers } = await prepareUserSession();
-            const response = await global.superAgent
-                .get(`${getUri(Constants.Endpoints.USERS)}/${id}`)
-                .set(headers);
-            expect(response.body.data.verified).toEqual(false);
-        });
+        // test('should be unverified by default', async () => {
+        //     const { body: { data: { id } } } = await userFixture.createUser({ verified: null });
+        //     const { headers } = await prepareUserSession();
+        //     const response = await global.superAgent
+        //         .get(`${getUri(Constants.Endpoints.USERS)}/${id}`)
+        //         .set(headers);
+        //     expect(response.body.data.verified).toEqual(false);
+        // });
 
         test.todo('user should have a defualt profile equal to empty {}');
     });
 
-    describe('#SEARCH FOR USES', () => {
-        test.skip('should return list of users depends on the provided name', () => {
-            global.superAgent.get(USER_SESSION_ENDPOINT);
-        });
-    });
+    // describe('#SEARCH FOR USES', () => {
+    //     test.skip('should return list of users depends on the provided name', () => {
+    //         global.superAgent.get(USER_SESSION_ENDPOINT);
+    //     });
+    // });
 
 });
