--- src/app/core/config/index.ts
@@ -1,3 +1,0 @@
-// export * from './config';
-export * from './observer';
-export * from './reactor';
