+++ src/app/core/response/response.ts
@@ -1,0 +1,9 @@
+import { NetworkStatus } from '@core/helpers/network-status';
+
+export abstract class HttpResponse extends Error {
+    public status: string;
+    constructor(public code: number) {
+        super();
+        this.status = NetworkStatus.getStatusText(code);
+    }
+}
