+++ src/app/core/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './utils.service';
 export * from './logger.service';
-export * from './performance.service';
+export * from './stopwatch.service';
