+++ src/environment/.env.development
@@ -3,4 +3,3 @@ MONGO_PASSWORD=27017
 MONGO_PATH=dev
 JWT_SECRET_KEY=secret
 NODE_STAGE=development
-CLIENT_RESETPASSWORD_URL=http://localhost:4200/portal/resetpassword/
