+++ src/environment/.env.production
@@ -1,2 +1,1 @@
 NODE_STAGE=production
-CLIENT_RESETPASSWORD_URL=https://angular-buildozer.herokuapp.com/portal/resetpassword
